+++ CHANGES.md
@@ -6,6 +6,8 @@ Fixed:
   (#2782)
 * Space trim in interactive variables set on telnet (#2785)
 * Fixed internal streaming logic in `max_duration` and `crossfade`.
+* Make sure that there's at most one metadata at any given
+  frame position (#2786)
 
 ---
 2.1.3 (2022-11-04)